Tiled, Aligned Memory Access

+++ viewshed.h
@@ -34,7 +34,7 @@
 #define __VIEWSHED_H__
 
 extern void viewshed(const float * src, float * dst,
-		     uint64_t cols, uint64_t rows,
+		     uint32_t cols, uint32_t rows,
 		     double xres, double yres);
 
 #endif
